refactor(cl): moved SimpleUserDataMixin to tests/mixins.py
